perf(tables): make `b-th` extend `b-td`  instead of using functional wrappers (#4156)
